WIP: Allow to execute a script on the orchestrator. #139
